requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.8.0
    # via flask-migrate
build==0.8.0
    # via pip-tools
click==8.1.3
    # via
    #   flask
    #   pip-tools
colorama==0.4.5
    # via
    #   build
    #   click
flask==2.1.2
    # via
    #   -r requirements.in
    #   flask-login
    #   flask-migrate
    #   flask-sqlalchemy
    #   flask-wtf
flask-login==0.6.1
    # via -r requirements.in
flask-migrate==3.1.0
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-migrate
flask-wtf==1.0.1
    # via -r requirements.in
greenlet==1.1.2
    # via sqlalchemy
importlib-metadata==4.12.0
    # via flask
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via flask
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
    #   wtforms
packaging==21.3
    # via build
pep517==0.12.0
    # via build
pip-tools==6.7.0
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
python-dotenv==0.20.0
    # via -r requirements.in
sqlalchemy==1.4.39
    # via
    #   alembic
    #   flask-sqlalchemy
tomli==2.0.1
    # via
    #   build
    #   pep517
werkzeug==2.1.2
    # via
    #   flask
    #   flask-login
wheel==0.37.1
    # via pip-tools
wtforms==3.0.1
    # via flask-wtf
zipp==3.8.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
